DEV-28: Update Documentation (#232)
* Updating touchpoint content

* small typo cleanup

* fix broken theme link

* fix formatting

* add language to chat mode examples
