--- packages/website/src/content/02-05-touchpoint-ui-api-reference.tsx
@@ -1,11 +1,0 @@
-import { type FC } from "react";
-import { PageContent } from "../components/PageContent";
-import content from "./touchpoint-ui-api-reference.md?raw";
-
-export const navGroup: string = "Touchpoint";
-
-export const title: string = "API reference";
-
-export const Content: FC<unknown> = () => {
-  return <PageContent md={content} />;
-};
