--- packages/website/src/content/03-04-touchpoint-CustomCards.tsx
@@ -1,14 +1,0 @@
-import { type FC } from "react";
-
-import { PageContent } from "../components/PageContent";
-import customCardContent from "./03-04-touchpoint-CustomCards.md?raw";
-
-export const content = customCardContent;
-
-export const navGroup: string = "Touchpoint components";
-
-export const title: string = "Custom Cards";
-
-export const Content: FC<unknown> = () => {
-  return <PageContent md={content} />;
-};
