--- packages/website/src/content/03-05-touchpoint-Icons.md
@@ -1,94 +1,0 @@
-- [About](#about)
-- [Icon Properties](#icon-properties)
-- [Import and Basic Usage](#import-and-basic-usage)
-- [Example](#example)
-- [Available Icons](#available-icons)
-- [Related Components](#related-components)
-
-## About
-
-Touchpoint provides built-in icons that integrate with buttons and custom components. Each icon maintains consistent styling and accessibility features.
-
-## Icon Properties
-
-Each icon accepts these properties:
-
-| Property  | Type   | Description                       |
-| --------- | ------ | --------------------------------- |
-| size      | number | Icon size in pixels (default: 24) |
-| color     | string | Icon color (inherits from parent) |
-| className | string | Additional CSS classes            |
-
-## Import and Basic Usage
-
-You can import the Icons from touchpoint once the package has been installed or made available in your project.
-
-The following examples use the Icons available from Touchpoint to construct an [IconButton](/touchpoint-Buttons).
-
-## Example
-
-Example IconButton using the Icons from Touchpoint
-
-```javascript
-import { useTouchpointContext, IconButton, Icons } from "@nlxai/touchpoint-ui";
-const IconButtonExample = ({ data, conversationHandler }) => {
-  return (
-    <IconButton
-      label={data.buttonLabel}
-      Icon={Icons.ArrowForward}
-      onClick={() => conversationHandler.sendChoice(data.buttonId)}
-      type="main"
-    />
-  );
-};
-```
-
-## Available Icons
-
-- Action
-- Assistant
-- AssistantOld
-- Add
-- ArrowDown
-- ArrowLeft
-- ArrowRight
-- ArrowUp
-- ArrowForward
-- Attachment
-- Camera
-- Check
-- Close
-- Copy
-- Date
-- Delete
-- Escalate
-- Error
-- FullScreen
-- Mic
-- MicOff
-- Location
-- Volume
-- VolumeOff
-- Translate
-- OpenInNew
-- Play
-- Preview
-- Reorder
-- Restart
-- Settings
-- Search
-- Share
-- Warning
-- ThumbDown
-- ThumbUp
-- Time
-- Undo
-- Refresh
-- Help
-
-## Related Components
-
-- [Buttons](/touchpoint-Buttons)
-- [CustomCards](/touchpoint-CustomCards)
-- [Theming Touchpoint](/touchpoint-ui-theming)
-- [Building Components without JSX](/touchpoint-components#building-html-components-without-transpiling)
