+++ packages/website/src/content/04-06-touchpoint-DateInput.tsx
@@ -1,0 +1,14 @@
+import { type FC } from "react";
+import dateContent from "./04-06-touchpoint-DateInput.md?raw";
+
+import { PageContent } from "../components/PageContent";
+
+export const content = dateContent;
+
+export const navGroup: string = "Touchpoint components";
+
+export const title: string = "DateInput";
+
+export const Content: FC<unknown> = () => {
+  return <PageContent md={content} />;
+};
